oh-tab-b7qq: extract secret commands
